requirements.txt
PyGithub==1.55
requests==2.26.0
loguru==0.5.3
